Update documentation (#25)
+ some refactring
+ some improves

 Changes to be committed:
	modified: .gitignore
	modified: README.md
	renamed: docs/concepts/concepts.md
	new file: docs/concepts/usecases.md
	new file: docs/gen_ref_pages.py
	modified: docs/index.md
	renamed: docs/notebooks/basic_example.ipynb
	new file: docs/notebooks/using_where.ipynb
	new file: docs/python/main_structures.md
	modified: mkdocs.yml
	new file: static/protobuf/protobuf-diagram.svg
	new file: static/protobuf/protobuf-diagram.uml
	modified: tsumugi-python/poetry.lock
	modified: tsumugi-python/pyproject.toml
	modified: tsumugi-python/tsumugi/__init__.py
	modified: tsumugi-python/tsumugi/analyzers.py
	modified: tsumugi-python/tsumugi/anomaly_detection.py
	modified: tsumugi-python/tsumugi/checks.py
	deleted: tsumugi-python/tsumugi/examples/base_example.py
	deleted: tsumugi-python/tsumugi/examples/base_example_classic.py
	modified: tsumugi-python/tsumugi/verification.py

+++ tsumugi-python/tsumugi/checks.py
@@ -44,7 +44,7 @@ class CheckBuilder:
         return self
 
     def has_size(
-        self, expected_size: float, hint: str = "", name: str | None = None
+        self, expected_size: int, hint: str = "", name: str | None = None
     ) -> Self:
         """Add a constraint that the DataFrame has size like expected."""
         return self.with_constraint(
